requirements.txt
Flask==3.0.0
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.97.0